Move validate function to location utils
