+++ app/utils/locations/index.ts
@@ -1,8 +1,8 @@
 import type { DataCentersList } from './DataCenters'
-import DataCenters from './DataCenters'
+import DataCenters, { DataCenterArray } from './DataCenters'
 import Regions from './Regions'
 import type { WorldsList } from './Worlds'
-import Worlds from './Worlds'
+import Worlds, { WorldsArray } from './Worlds'
 import {
   DataCenterNotFoundException,
   RegionNotFoundException
@@ -29,4 +29,24 @@ const WorldsOfDataCenter: (
   throw new DataCenterNotFoundException(data_center_key)
 }
 
-export { Regions, Worlds, DataCenters, DataCentersOfRegion, WorldsOfDataCenter }
+const validateWorldAndDataCenter = (
+  world?: string | null,
+  data_center?: string | null
+) => {
+  if (world && WorldsArray.includes(world)) {
+    if (data_center && DataCenterArray.includes(data_center)) {
+      return { world, data_center }
+    }
+  }
+
+  return { world: WorldsArray.at(0), data_center: DataCenterArray.at(0) }
+}
+
+export {
+  Regions,
+  Worlds,
+  DataCenters,
+  DataCentersOfRegion,
+  WorldsOfDataCenter,
+  validateWorldAndDataCenter
+}
